Fixed Flavor Mapping Bug for VM Size Update
Summary: Whenever the VM size is changed via platform, in case if the node already contains a value for 'vm_size', that value gets picked instead of the new size value. That bug has been fixed in this code to check if there is a size update, then always pick the latest flavor for the updated size.