Skip Azure Create Call in Case of Touch/Update
Summary: Whenever an update is run on a compute in OneOps, it calls the 'azure::add_node.rb' recipe which can end up changing the IP address if the IP address is DYNAMIC. Here we have handled the case that in case of update, do not make any create_or_update calls, rather skip those and move on. The case for accelerated networking has been handled as well in case a user wants to turn it off or on.